Revert "[backport] Use directional derivative"

+++ tests/chainer_tests/test_gradient_check.py
@@ -354,7 +354,7 @@ class TestCheckBackward(unittest.TestCase):
             s = Ident()(x)
             return s,
 
-        gradient_check.check_backward(f, (x1, x2), g1, dtype=self.dtype)
+        gradient_check.check_backward(f, (x1, x2), g1)
 
     def test_no_grads_option(self):
         x1 = numpy.array([1], dtype='f')
